Minimize rebuilds of RAFT code due to knowhere changes
