+++ src/common/raft/integration/ivf_pq_instantiations.cu
@@ -15,16 +15,13 @@
  * limitations under the License.
  */
 #undef RAFT_EXPLICIT_INSTANTIATE_ONLY
-#include "common/raft/integration/raft_knowhere_index.cuh"
+#include "common/raft/integration/type_mappers.hpp"
 #include "common/raft/proto/filtered_search_instantiation.cuh"
 #include "common/raft/proto/raft_index_kind.hpp"
 
-RAFT_FILTERED_SEARCH_EXTERN(ivf_flat, raft_knowhere::raft_data_t<raft_proto::raft_index_kind::ivf_flat>,
-                            raft_knowhere::raft_indexing_t<raft_proto::raft_index_kind::ivf_flat>,
-                            raft_knowhere::raft_input_indexing_t<raft_proto::raft_index_kind::ivf_flat>,
-                            raft_knowhere::raft_data_t<raft_proto::raft_index_kind::ivf_flat>,
-                            raft_knowhere::knowhere_bitset_data_type, raft_knowhere::knowhere_bitset_indexing_type)
-
-namespace raft_knowhere {
-template struct raft_knowhere_index<raft_proto::raft_index_kind::ivf_flat>;
-}  // namespace raft_knowhere
+RAFT_FILTERED_SEARCH_INSTANTIATION(ivf_pq, raft_knowhere::raft_data_t<raft_proto::raft_index_kind::ivf_pq>,
+                                   raft_knowhere::raft_indexing_t<raft_proto::raft_index_kind::ivf_pq>,
+                                   raft_knowhere::raft_input_indexing_t<raft_proto::raft_index_kind::ivf_pq>,
+                                   raft_knowhere::raft_data_t<raft_proto::raft_index_kind::ivf_pq>,
+                                   raft_knowhere::knowhere_bitset_data_type,
+                                   raft_knowhere::knowhere_bitset_indexing_type)
