+++ src/common/raft/integration/raft_knowhere_index.hpp
@@ -18,51 +18,10 @@
 #include <cstdint>
 
 #include "common/raft/integration/raft_knowhere_config.hpp"
+#include "common/raft/integration/type_mappers.hpp"
 #include "common/raft/proto/raft_index_kind.hpp"
 
 namespace raft_knowhere {
-
-using knowhere_data_type = float;
-using knowhere_indexing_type = std::int64_t;
-using knowhere_bitset_data_type = std::uint8_t;
-using knowhere_bitset_indexing_type = std::uint32_t;
-
-namespace detail {
-
-template <bool B, raft_proto::raft_index_kind IndexKind>
-struct raft_io_type_mapper : std::false_type {};
-
-template <>
-struct raft_io_type_mapper<true, raft_proto::raft_index_kind::ivf_flat> : std::true_type {
-    using data_type = float;
-    using indexing_type = std::int64_t;
-    using input_indexing_type = std::int64_t;
-};
-
-template <>
-struct raft_io_type_mapper<true, raft_proto::raft_index_kind::ivf_pq> : std::true_type {
-    using data_type = float;
-    using indexing_type = std::int64_t;
-    using input_indexing_type = std::int64_t;
-};
-
-template <>
-struct raft_io_type_mapper<true, raft_proto::raft_index_kind::cagra> : std::true_type {
-    using data_type = float;
-    using indexing_type = std::uint32_t;
-    using input_indexing_type = std::int64_t;
-};
-
-}  // namespace detail
-
-template <raft_proto::raft_index_kind IndexKind>
-using raft_data_t = typename detail::raft_io_type_mapper<true, IndexKind>::data_type;
-
-template <raft_proto::raft_index_kind IndexKind>
-using raft_indexing_t = typename detail::raft_io_type_mapper<true, IndexKind>::indexing_type;
-
-template <raft_proto::raft_index_kind IndexKind>
-using raft_input_indexing_t = typename detail::raft_io_type_mapper<true, IndexKind>::input_indexing_type;
 
 template <raft_proto::raft_index_kind IndexKind>
 struct raft_knowhere_index {
